hwmon: ina3221: use s32 type for tables used with find_closest()
This is part of a rework for the find_closest() macro to become a function.
The intent is to make all arrays (used for find_closest() to be signed
32-bit. 'int' is 32-bit already, so this change is mostly cosmetic for the
'ina3221_avg_samples' table.

For the 'ina3221_conv_time' table, the type is converted from u16 to s32.

+++ drivers/hwmon/ina3221.c
@@ -178,12 +178,12 @@ static inline int ina3221_summation_shunt_resistor(struct ina3221_data *ina)
 }
 
 /* Lookup table for Bus and Shunt conversion times in usec */
-static const u16 ina3221_conv_time[] = {
+static const s32 ina3221_conv_time[] = {
 	140, 204, 332, 588, 1100, 2116, 4156, 8244,
 };
 
 /* Lookup table for number of samples using in averaging mode */
-static const int ina3221_avg_samples[] = {
+static const s32 ina3221_avg_samples[] = {
 	1, 4, 16, 64, 128, 256, 512, 1024,
 };
